+++ internal/module/expense/entity/expense_participant_entity.go
@@ -6,6 +6,6 @@ type ExpenseParticipant struct {
 	ID            string
 	ExpenseID     string
 	ParticipantID string
-	ShareAmount   float64
+	ShareAmount   int64
 	CreatedAt     time.Time
 }
